Run the test suite against multiple databases

+++ test/dummy/db/schema.rb
@@ -81,9 +81,9 @@ ActiveRecord::Schema[8.1].define(version: 2026_08_09_125149) do
   end
 
   create_table "roles", force: :cascade do |t|
-    t.integer "company_id", null: false
+    t.bigint "company_id", null: false
     t.datetime "created_at", null: false
-    t.integer "person_id", null: false
+    t.bigint "person_id", null: false
     t.string "title"
     t.datetime "updated_at", null: false
     t.index ["company_id"], name: "index_roles_on_company_id"
@@ -91,7 +91,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_08_09_125149) do
   end
 
   create_table "titles", force: :cascade do |t|
-    t.integer "book_id", null: false
+    t.bigint "book_id", null: false
     t.datetime "created_at", null: false
     t.string "locale"
     t.string "title"
